+++ CHANGELOG.md
@@ -435,10 +435,12 @@
   - [v0.2.0](services/telemetryrouter/CHANGELOG.md#v020)
     - **Feature:** Added `_UNKNOWN_DEFAULT_OPEN_API` fallback value to all enums to handle unknown API values gracefully.
   - [v0.2.1](services/telemetryrouter/CHANGELOG.md#v021)
-    - **Improvement**: Use new `WaiterHandler` struct in the DNS WaitHandler
+    - **Improvement**: Use new `WaiterHandler` struct in the TelemetryRouter WaitHandler
 - `telemetrylink`:
   - [v0.1.0](services/telemetrylink/CHANGELOG.md#v010)
     - **New**: API for STACKIT Telemetry Link
+  - [v0.1.1](services/telemetryrouter/CHANGELOG.md#v021)
+    - **Improvement**: Use new `WaiterHandler` struct in the TelemetryLink WaitHandler
 - `vpn`:
   - [v0.4.2](services/vpn/CHANGELOG.md#v042) 
     - **Dependencies:** Bump STACKIT SDK core module from `v0.24.0` to `v0.24.1`